requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
exceptiongroup==1.0.0
glob2==0.7
h11==0.14.0
idna==3.4
iniconfig==1.1.1
load-dotenv==0.1.0
Mako==1.2.3
MarkupSafe==2.1.1
outcome==1.2.0
packaging==21.3
parse==1.19.0
parse-type==0.6.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.0
pytest-bdd==6.0.1
pytest-html==3.2.0
pytest-metadata==2.0.4
python-dotenv==0.21.0
requests==2.28.1
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
